+++ record_ios/ios/record_ios/Sources/record_ios/delegate/RecorderDelegateProtocols.swift
@@ -12,6 +12,8 @@ public enum RecordState: Int {
 }
 
 public protocol AudioRecordingDelegate {
+  var config: RecordConfig? { get }
+  
   func stop(completionHandler: @escaping (String?) -> ())
   func cancel() throws
   func getAmplitude() -> Float