+++ record_platform_interface/lib/src/types/record_config.dart
@@ -60,6 +60,21 @@ class RecordConfig {
   /// iOS specific configuration.
   final IosRecordConfig iosConfig;
 
+  /// Recorder behaviour when audio is interrupted by another source.
+  ///
+  /// System alerts are ignored.
+  /// Some other sources may not be detected (e.g. browser).
+  ///
+  /// Platforms: Android & iOS.
+  final AudioInterruptionMode audioInterruption;
+
+  /// Useful for those who need finer data when streaming.
+  ///
+  /// Underlying implementations may adjust to other value or throw exception if under miminum size required.
+  ///
+  /// Platforms: Android, iOS, macOS & web.
+  final int? streamBufferSize;
+
   const RecordConfig({
     this.encoder = AudioEncoder.aacLc,
     this.bitRate = 128000,
@@ -71,6 +86,8 @@ class RecordConfig {
     this.noiseSuppress = false,
     this.androidConfig = const AndroidRecordConfig(),
     this.iosConfig = const IosRecordConfig(),
+    this.audioInterruption = AudioInterruptionMode.pause,
+    this.streamBufferSize,
   });
 
   Map<String, dynamic> toMap() {
@@ -85,6 +102,8 @@ class RecordConfig {
       'noiseSuppress': noiseSuppress,
       'androidConfig': androidConfig.toMap(),
       'iosConfig': iosConfig.toMap(),
+      'audioInterruption': audioInterruption.index,
+      'streamBufferSize': streamBufferSize,
     };
   }
 }