+++ record_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.4.0
+* feat: Add AudioInterruptionMode to `RecordConfig`.
+* feat: Add stream buffer size option.
+
 ## 1.3.0
 * feat: Add `audioManagerMode` and `speakerphone` options to `AndroidRecordConfig`.